requirements.txt
amqp==1.4.9
dj-database-url==0.4.1
Django==1.10.2
django-cors-headers==1.2.2
django-cors-middleware==1.3.1
django-oauth-toolkit==0.11.0
djangorestframework==3.5.4
future==0.16.0
gunicorn==19.6.0
python-dateutil==2.5.3
pytz==2016.7
whitenoise==3.2.2
psycopg2
